refactor: share coin form for harvest
